requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==3.0.6
WebTest==2.0.25
factory-boy==2.8.1

# Lint and code style
flake8==3.3.0
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
flake8-docstrings==1.0.3
flake8-isort==2.1.3
flake8-quotes==0.9.0
isort==4.2.5
pep8-naming==0.4.1




#Mail
flask-mail

# chat
chatterbot
apiai

# files
Flask-Excel
pyexcel-xls